Fix TUI scroll overflow: use u32 for scroll arithmetic
The scroll_offset, last_view_height, and content_height were all u16,
which silently overflows when wrapped content exceeds 65535 lines. This
caused long responses to appear to flush previous conversation history
because the scroll position wrapped around to near-zero.

Widen all scroll arithmetic to u32 and only clamp to u16 at the final
ratatui scroll() call.

+++ crates/quine-cli/src/tui/app.rs
@@ -307,7 +307,7 @@ impl fmt::Display for InputBuffer {
 pub struct App {
     pub messages: Vec<ConversationEntry>,
     pub streaming_buffer: String,
-    pub scroll_offset: u16,
+    pub scroll_offset: u32,
     pub user_scrolled: bool,
     pub input: InputBuffer,
     pub interaction_queue: VecDeque<PendingInteraction>,
@@ -327,7 +327,7 @@ pub struct App {
     /// Whether this session is in read-only plan mode.
     pub plan_mode: bool,
     /// Last known conversation view height (set during rendering for scroll step sizing).
-    pub last_view_height: u16,
+    pub last_view_height: u32,
 }
 
 impl App {
@@ -834,12 +834,12 @@ impl App {
         }
     }
 
-    pub fn scroll_up(&mut self, amount: u16) {
+    pub fn scroll_up(&mut self, amount: u32) {
         self.scroll_offset = self.scroll_offset.saturating_add(amount);
         self.user_scrolled = true;
     }
 
-    pub fn scroll_down(&mut self, amount: u16) {
+    pub fn scroll_down(&mut self, amount: u32) {
         self.scroll_offset = self.scroll_offset.saturating_sub(amount);
         if self.scroll_offset == 0 {
             self.user_scrolled = false;
